factor out fn for ghash prep
