+++ aead/src/aes_gcm/mod.rs
@@ -54,6 +54,29 @@ impl Add for AesGcmTagShare {
     }
 }
 
+/// Builds padded data for GHASH
+pub(crate) fn build_ghash_data(mut aad: Vec<u8>, mut ciphertext: Vec<u8>) -> Vec<u8> {
+    let associated_data_bitlen = (aad.len() as u64) * 8;
+    let text_bitlen = (ciphertext.len() as u64) * 8;
+
+    let len_block = ((associated_data_bitlen as u128) << 64) + (text_bitlen as u128);
+
+    // pad data to be a multiple of 16 bytes
+    let aad_padded_block_count = (aad.len() / 16) + (aad.len() % 16 != 0) as usize;
+    aad.resize(aad_padded_block_count * 16, 0);
+
+    let ciphertext_padded_block_count =
+        (ciphertext.len() / 16) + (ciphertext.len() % 16 != 0) as usize;
+    ciphertext.resize(ciphertext_padded_block_count * 16, 0);
+
+    let mut data: Vec<u8> = Vec::with_capacity(aad.len() + ciphertext.len() + 8);
+    data.extend(aad);
+    data.extend(ciphertext);
+    data.extend_from_slice(&len_block.to_be_bytes());
+
+    data
+}
+
 #[cfg(test)]
 mod tests {
     use crate::{AeadFollower, AeadLeader};
